+++ src/openrct2/world/Park.h
@@ -90,7 +90,7 @@ namespace OpenRCT2
 extern uint64_t gParkFlags;
 extern uint16_t gParkRating;
 extern money16 gParkEntranceFee;
-extern uint32_t gParkSize;
+extern uint16_t gParkSize;
 extern money16 gLandPrice;
 extern money16 gConstructionRightsPrice;
